Introducing TextureMap Mask.
